chore: rename consolidated and move tests
Ref. metriport/metriport-internal#870
