+++ packages/api/src/command/medical/patient/__tests__/consolidated-get.test.ts
@@ -3,10 +3,10 @@ import * as dotenv from "dotenv";
 dotenv.config({ path: ".env.test" });
 import { Observation } from "@medplum/fhirtypes";
 import { v4 as uuidv4 } from "uuid";
-import { HapiFhirClient } from "../../../external/fhir/api/api-hapi";
-import { makePatient } from "../../../models/medical/__tests__/patient";
-import { getConsolidatedPatientData } from "../patient/get-consolidate-data";
-import * as getPatient from "../patient/get-patient";
+import { HapiFhirClient } from "../../../../external/fhir/api/api-hapi";
+import { makePatient } from "../../../../models/medical/__tests__/patient";
+import { getConsolidatedPatientData } from "../consolidated-get";
+import * as getPatient from "../get-patient";
 
 let getPatientOrFailMock: jest.SpyInstance;
 let fhir_searchResourcePages: jest.SpyInstance;
