+++ packages/api/src/command/medical/patient/consolidated-create.ts
