+++ packages/api/src/command/medical/patient/consolidated-get.ts
